+++ src/lib/util/entities.js
@@ -1,0 +1,12 @@
+"use strict";
+
+exports.NBSP = "\u00A0";
+exports.RSQUO = "\u2019";
+exports.LSQUO = "\u2018";
+exports.RDQUO = "\u201D";
+exports.LDQUO = "\u201C";
+exports.EMDASH = "\u2014";
+exports.ENDASH = "\u2013";
+exports.REG = "\u00AE";
+exports.TM = "\u2122";
+exports.BULL = "\u2022";